--- brainsets/brainsets/processing/__init__.py
@@ -1,14 +1,0 @@
-from .signal import downsample_wideband, extract_bands, cube_to_long
-
-__all__ = [
-    "downsample_wideband",
-    "extract_bands",
-    "cube_to_long",
-]
-
-
-# Drives the generated API reference; see docs/source/api_reference.py.
-__api_ref__ = {
-    "description": None,
-    "sections": [{"autosummary": __all__}],
-}
