--- brainsets/brainsets/utils/__init__.py
@@ -1,7 +1,0 @@
-from . import dandi_utils
-from . import split
-from . import bids_utils
-from . import mne_utils
-from . import s3_utils
-from . import openneuro
-from . import misc_utils
