+++ tests/utils/test_bids.py
@@ -16,7 +16,7 @@ except ImportError:
     MNE_BIDS_AVAILABLE = False
 
 try:
-    from brainsets.utils.bids_utils import (
+    from torch_brain.utils.bids import (
         EEG_EXTENSIONS,
         IEEG_EXTENSIONS,
         fetch_eeg_recordings,
@@ -46,7 +46,7 @@ except ImportError:
 
 
 # ============================================================================
-# BIDS-Specific Fixtures (for test_bids_utils.py only)
+# BIDS-Specific Fixtures (for test_bids.py only)
 # ============================================================================
 @pytest.fixture
 def bids_root(tmp_path):
@@ -1112,19 +1112,19 @@ def test_eeg_extensions_do_not_include_nwb():
 class TestCheckMneBidsAvailable:
     """Test that functions raise ImportError when MNE_BIDS is not available."""
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_fetch_eeg_recordings_raises_import_error(self):
         """Test that fetch_eeg_recordings raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import fetch_eeg_recordings
+        from torch_brain.utils.bids import fetch_eeg_recordings
 
         source = MagicMock()
         with pytest.raises(ImportError, match="fetch_eeg_recordings requires mne-bids"):
             fetch_eeg_recordings(source)
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_fetch_ieeg_recordings_raises_import_error(self):
         """Test that fetch_ieeg_recordings raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import fetch_ieeg_recordings
+        from torch_brain.utils.bids import fetch_ieeg_recordings
 
         source = MagicMock()
         with pytest.raises(
@@ -1132,10 +1132,10 @@ class TestCheckMneBidsAvailable:
         ):
             fetch_ieeg_recordings(source)
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_group_recordings_by_entity_raises_import_error(self):
         """Test that group_recordings_by_entity raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import group_recordings_by_entity
+        from torch_brain.utils.bids import group_recordings_by_entity
 
         recordings = MagicMock()
         with pytest.raises(
@@ -1143,10 +1143,10 @@ class TestCheckMneBidsAvailable:
         ):
             group_recordings_by_entity(recordings)
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_check_eeg_recording_files_exist_raises_import_error(self):
         """Test that check_eeg_recording_files_exist raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import check_eeg_recording_files_exist
+        from torch_brain.utils.bids import check_eeg_recording_files_exist
 
         bids_root = MagicMock()
         recording_id = MagicMock()
@@ -1155,10 +1155,10 @@ class TestCheckMneBidsAvailable:
         ):
             check_eeg_recording_files_exist(bids_root, recording_id)
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_check_ieeg_recording_files_exist_raises_import_error(self):
         """Test that check_ieeg_recording_files_exist raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import check_ieeg_recording_files_exist
+        from torch_brain.utils.bids import check_ieeg_recording_files_exist
 
         bids_root = MagicMock()
         recording_id = MagicMock()
@@ -1167,19 +1167,19 @@ class TestCheckMneBidsAvailable:
         ):
             check_ieeg_recording_files_exist(bids_root, recording_id)
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_load_json_sidecar_raises_import_error(self):
         """Test that load_json_sidecar raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import load_json_sidecar
+        from torch_brain.utils.bids import load_json_sidecar
 
         bids_path = MagicMock()
         with pytest.raises(ImportError, match="load_json_sidecar requires mne-bids"):
             load_json_sidecar(bids_path)
 
-    @patch("brainsets.utils.bids_utils.MNE_BIDS_AVAILABLE", False)
+    @patch("torch_brain.utils.bids.MNE_BIDS_AVAILABLE", False)
     def test_build_bids_path_raises_import_error(self):
         """Test that build_bids_path raises ImportError when MNE_BIDS is unavailable."""
-        from brainsets.utils.bids_utils import build_bids_path
+        from torch_brain.utils.bids import build_bids_path
 
         bids_root = MagicMock()
         recording_id = MagicMock()
